Revert "Move metadata specific methods to ElasticsearchMetadata"
This reverts commit 335af1ab16c366afffa97234fced91d8bd6ed467.
