+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/ElasticsearchConnectorModule.java
@@ -20,9 +20,13 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.FromStringDeserializer;
 import com.google.inject.Binder;
 import com.google.inject.Module;
+import com.google.inject.Provides;
 import com.google.inject.Scopes;
 
 import javax.inject.Inject;
+import javax.inject.Singleton;
+
+import java.io.IOException;
 
 import static com.facebook.airlift.configuration.ConfigBinder.configBinder;
 import static com.facebook.airlift.json.JsonBinder.jsonBinder;
@@ -41,7 +45,6 @@ public class ElasticsearchConnectorModule
         binder.bind(ElasticsearchMetadata.class).in(Scopes.SINGLETON);
         binder.bind(ElasticsearchSplitManager.class).in(Scopes.SINGLETON);
         binder.bind(ElasticsearchRecordSetProvider.class).in(Scopes.SINGLETON);
-        binder.bind(ElasticsearchClient.class).in(Scopes.SINGLETON);
 
         configBinder(binder).bindConfig(ElasticsearchConfig.class);
 
@@ -49,6 +52,14 @@ public class ElasticsearchConnectorModule
         jsonCodecBinder(binder).bindJsonCodec(ElasticsearchTableDescription.class);
 
         binder.install(new DecoderModule());
+    }
+
+    @Singleton
+    @Provides
+    public static ElasticsearchClient createElasticsearchClient(ElasticsearchConfig config, ElasticsearchTableDescriptionProvider elasticsearchTableDescriptionProvider)
+            throws IOException
+    {
+        return new ElasticsearchClient(elasticsearchTableDescriptionProvider, config);
     }
 
     private static final class TypeDeserializer
