+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/ElasticsearchMetadata.java
@@ -13,10 +13,6 @@
  */
 package com.facebook.presto.elasticsearch;
 
-import com.facebook.airlift.json.ObjectMapperProvider;
-import com.facebook.airlift.log.Logger;
-import com.facebook.presto.common.type.RowType;
-import com.facebook.presto.common.type.Type;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ColumnMetadata;
 import com.facebook.presto.spi.ConnectorSession;
@@ -26,98 +22,44 @@ import com.facebook.presto.spi.ConnectorTableLayoutHandle;
 import com.facebook.presto.spi.ConnectorTableLayoutResult;
 import com.facebook.presto.spi.ConnectorTableMetadata;
 import com.facebook.presto.spi.Constraint;
-import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.SchemaTablePrefix;
 import com.facebook.presto.spi.TableNotFoundException;
 import com.facebook.presto.spi.connector.ConnectorMetadata;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.util.concurrent.UncheckedExecutionException;
-import org.elasticsearch.cluster.metadata.MappingMetaData;
-import org.elasticsearch.common.collect.ImmutableOpenMap;
 
-import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.TreeMap;
-import java.util.concurrent.ExecutorService;
 
-import static com.facebook.airlift.concurrent.Threads.daemonThreadsNamed;
-import static com.facebook.presto.common.type.BigintType.BIGINT;
-import static com.facebook.presto.common.type.BooleanType.BOOLEAN;
-import static com.facebook.presto.common.type.DoubleType.DOUBLE;
-import static com.facebook.presto.common.type.IntegerType.INTEGER;
-import static com.facebook.presto.common.type.RowType.Field;
-import static com.facebook.presto.common.type.VarbinaryType.VARBINARY;
-import static com.facebook.presto.common.type.VarcharType.VARCHAR;
-import static com.facebook.presto.elasticsearch.ElasticsearchErrorCode.ELASTICSEARCH_CORRUPTED_MAPPING_METADATA;
-import static com.google.common.base.Throwables.throwIfInstanceOf;
-import static com.google.common.cache.CacheLoader.asyncReloading;
-import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.Executors.newFixedThreadPool;
-import static java.util.concurrent.TimeUnit.MINUTES;
 
 public class ElasticsearchMetadata
         implements ConnectorMetadata
 {
-    private static final Logger LOG = Logger.get(ElasticsearchMetadata.class);
-    private final ObjectMapper objectMapper = new ObjectMapperProvider().get();
-
-    private final ElasticsearchTableDescriptionProvider tableDescriptions;
-    private final LoadingCache<ElasticsearchTableDescription, List<ColumnMetadata>> columnMetadataCache;
-    private final ExecutorService executor = newFixedThreadPool(1, daemonThreadsNamed("elasticsearch-metadata-%s"));
     private final ElasticsearchClient client;
 
     @Inject
-    public ElasticsearchMetadata(ElasticsearchClient client, ElasticsearchTableDescriptionProvider descriptions)
+    public ElasticsearchMetadata(ElasticsearchClient client)
     {
         this.client = requireNonNull(client, "client is null");
-        this.tableDescriptions = requireNonNull(descriptions, "descriptions is null");
-
-        this.columnMetadataCache = CacheBuilder.newBuilder()
-                .expireAfterWrite(30, MINUTES)
-                .refreshAfterWrite(15, MINUTES)
-                .maximumSize(500)
-                .build(asyncReloading(CacheLoader.from(this::loadColumns), executor));
-    }
-
-    @PreDestroy
-    public void close()
-    {
-        executor.shutdown();
     }
 
     @Override
     public List<String> listSchemaNames(ConnectorSession session)
     {
-        return tableDescriptions.getAllSchemaTableNames()
-                .stream()
-                .map(SchemaTableName::getSchemaName)
-                .collect(toImmutableList());
+        return client.listSchemas();
     }
 
     @Override
     public ElasticsearchTableHandle getTableHandle(ConnectorSession session, SchemaTableName tableName)
     {
         requireNonNull(tableName, "tableName is null");
-        ElasticsearchTableDescription table = getTable(tableName.getSchemaName(), tableName.getTableName());
+        ElasticsearchTableDescription table = client.getTable(tableName.getSchemaName(), tableName.getTableName());
         if (table == null) {
             return null;
         }
@@ -150,10 +92,7 @@ public class ElasticsearchMetadata
     @Override
     public List<SchemaTableName> listTables(ConnectorSession session, Optional<String> schemaName)
     {
-        return tableDescriptions.getAllSchemaTableNames()
-                .stream()
-                .filter(schemaTableName -> !schemaName.isPresent() || schemaTableName.getSchemaName().equals(schemaName.get()))
-                .collect(toImmutableList());
+        return client.listTables(schemaName);
     }
 
     @Override
@@ -161,39 +100,28 @@ public class ElasticsearchMetadata
     {
         ElasticsearchTableHandle handle = (ElasticsearchTableHandle) tableHandle;
 
-        ElasticsearchTableDescription table = getTable(handle.getSchemaName(), handle.getTableName());
+        ElasticsearchTableDescription table = client.getTable(handle.getSchemaName(), handle.getTableName());
         if (table == null) {
             throw new TableNotFoundException(handle.getSchemaTableName());
         }
 
         ImmutableMap.Builder<String, ColumnHandle> columnHandles = ImmutableMap.builder();
         int index = 0;
-        for (ColumnMetadata column : getColumnMetadata(table)) {
+        for (ColumnMetadata column : client.getColumnMetadata(table)) {
             Map<String, Object> properties = column.getProperties();
             int ordinalPosition = (Integer) properties.get("ordinalPosition");
             int position = ordinalPosition == -1 ? index : ordinalPosition;
             columnHandles.put(column.getName(),
                     new ElasticsearchColumnHandle(
-                            String.valueOf(properties.get("originalColumnName")),
-                            column.getType(),
-                            String.valueOf(properties.get("jsonPath")),
-                            String.valueOf(properties.get("jsonType")),
-                            position,
-                            (Boolean) properties.get("isList")));
+                        String.valueOf(properties.get("originalColumnName")),
+                        column.getType(),
+                        String.valueOf(properties.get("jsonPath")),
+                        String.valueOf(properties.get("jsonType")),
+                        position,
+                        (Boolean) properties.get("isList")));
             index++;
         }
         return columnHandles.build();
-    }
-
-    private List<ColumnMetadata> getColumnMetadata(ElasticsearchTableDescription table)
-    {
-        try {
-            return columnMetadataCache.getUnchecked(table);
-        }
-        catch (UncheckedExecutionException e) {
-            throwIfInstanceOf(e.getCause(), PrestoException.class);
-            throw e;
-        }
     }
 
     @Override
@@ -219,11 +147,11 @@ public class ElasticsearchMetadata
 
     private Optional<ConnectorTableMetadata> getTableMetadata(SchemaTableName tableName)
     {
-        ElasticsearchTableDescription table = getTable(tableName.getSchemaName(), tableName.getTableName());
+        ElasticsearchTableDescription table = client.getTable(tableName.getSchemaName(), tableName.getTableName());
         if (table == null) {
             return Optional.empty();
         }
-        return Optional.of(new ConnectorTableMetadata(tableName, getColumnMetadata(table)));
+        return Optional.of(new ConnectorTableMetadata(tableName, client.getColumnMetadata(table)));
     }
 
     private List<SchemaTableName> listTables(ConnectorSession session, SchemaTablePrefix prefix)
@@ -232,227 +160,5 @@ public class ElasticsearchMetadata
             return listTables(session, Optional.empty());
         }
         return ImmutableList.of(new SchemaTableName(prefix.getSchemaName(), prefix.getTableName()));
-    }
-
-    private ElasticsearchTableDescription getTable(String schemaName, String tableName)
-    {
-        requireNonNull(schemaName, "schemaName is null");
-        requireNonNull(tableName, "tableName is null");
-        ElasticsearchTableDescription table = tableDescriptions.get(new SchemaTableName(schemaName, tableName));
-        if (table == null) {
-            return null;
-        }
-        if (table.getColumns().isPresent()) {
-            return table;
-        }
-        return new ElasticsearchTableDescription(
-                table.getTableName(),
-                table.getSchemaName(),
-                table.getIndex(),
-                table.getType(),
-                Optional.of(buildColumns(table)));
-    }
-
-    private List<ColumnMetadata> loadColumns(ElasticsearchTableDescription table)
-    {
-        if (table.getColumns().isPresent()) {
-            return buildMetadata(table.getColumns().get());
-        }
-        return buildMetadata(buildColumns(table));
-    }
-
-    private List<ColumnMetadata> buildMetadata(List<ElasticsearchColumn> columns)
-    {
-        List<ColumnMetadata> result = new ArrayList<>();
-        for (ElasticsearchColumn column : columns) {
-            Map<String, Object> properties = new HashMap<>();
-            properties.put("originalColumnName", column.getName());
-            properties.put("jsonPath", column.getJsonPath());
-            properties.put("isList", column.isList());
-            properties.put("ordinalPosition", column.getOrdinalPosition());
-            result.add(new ColumnMetadata(column.getName(), column.getType(), "", "", false, properties));
-        }
-        return result;
-    }
-
-    private List<ElasticsearchColumn> buildColumns(ElasticsearchTableDescription tableDescription)
-    {
-        ImmutableOpenMap<String, ImmutableOpenMap<String, MappingMetaData>> mappings = client.getMappings(tableDescription.getIndex(), tableDescription.getType());
-
-        List<ElasticsearchColumn> columns = new ArrayList<>();
-        Iterator<String> indexIterator = mappings.keysIt();
-        while (indexIterator.hasNext()) {
-            // TODO use io.airlift.json.JsonCodec
-            MappingMetaData mappingMetaData = mappings.get(indexIterator.next()).get(tableDescription.getType());
-            JsonNode rootNode;
-            try {
-                rootNode = objectMapper.readTree(mappingMetaData.source().uncompressed());
-            }
-            catch (IOException e) {
-                throw new PrestoException(ELASTICSEARCH_CORRUPTED_MAPPING_METADATA, e);
-            }
-            // parse field mapping JSON: https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-get-field-mapping.html
-            JsonNode mappingNode = rootNode.get(tableDescription.getType());
-            JsonNode propertiesNode = mappingNode.get("properties");
-
-            List<String> lists = new ArrayList<>();
-            JsonNode metaNode = mappingNode.get("_meta");
-            if (metaNode != null) {
-                JsonNode arrayNode = metaNode.get("lists");
-                if (arrayNode != null && arrayNode.isArray()) {
-                    ArrayNode arrays = (ArrayNode) arrayNode;
-                    for (int i = 0; i < arrays.size(); i++) {
-                        lists.add(arrays.get(i).textValue());
-                    }
-                }
-            }
-            populateColumns(propertiesNode, lists, columns);
-        }
-        return columns;
-    }
-
-    private void populateColumns(JsonNode propertiesNode, List<String> arrays, List<ElasticsearchColumn> columns)
-    {
-        FieldNestingComparator comparator = new FieldNestingComparator();
-        TreeMap<String, Type> fieldsMap = new TreeMap<>(comparator);
-        for (String columnMetadata : getColumnMetadata(Optional.empty(), propertiesNode)) {
-            int delimiterIndex = columnMetadata.lastIndexOf(":");
-            if (delimiterIndex == -1 || delimiterIndex == columnMetadata.length() - 1) {
-                LOG.debug("Invalid column path format: %s", columnMetadata);
-                continue;
-            }
-            String fieldName = columnMetadata.substring(0, delimiterIndex);
-            String typeName = columnMetadata.substring(delimiterIndex + 1);
-
-            if (!fieldName.endsWith(".type")) {
-                LOG.debug("Ignoring column with no type info: %s", columnMetadata);
-                continue;
-            }
-            String propertyName = fieldName.substring(0, fieldName.lastIndexOf('.'));
-            String nestedName = propertyName.replaceAll("properties\\.", "");
-            if (nestedName.contains(".")) {
-                fieldsMap.put(nestedName, getPrestoType(typeName));
-            }
-            else {
-                boolean newColumnFound = columns.stream()
-                        .noneMatch(column -> column.getName().equalsIgnoreCase(nestedName));
-                if (newColumnFound) {
-                    columns.add(new ElasticsearchColumn(nestedName, getPrestoType(typeName), nestedName, typeName, arrays.contains(nestedName), -1));
-                }
-            }
-        }
-        processNestedFields(fieldsMap, columns, arrays);
-    }
-
-    private void processNestedFields(TreeMap<String, Type> fieldsMap, List<ElasticsearchColumn> columns, List<String> arrays)
-    {
-        if (fieldsMap.size() == 0) {
-            return;
-        }
-        Map.Entry<String, Type> first = fieldsMap.firstEntry();
-        String field = first.getKey();
-        Type type = first.getValue();
-        if (field.contains(".")) {
-            String prefix = field.substring(0, field.lastIndexOf('.'));
-            ImmutableList.Builder<Field> fieldsBuilder = ImmutableList.builder();
-            int size = field.split("\\.").length;
-            Iterator<String> iterator = fieldsMap.navigableKeySet().iterator();
-            while (iterator.hasNext()) {
-                String name = iterator.next();
-                if (name.split("\\.").length == size && name.startsWith(prefix)) {
-                    Optional<String> columnName = Optional.of(name.substring(name.lastIndexOf('.') + 1));
-                    Type columnType = fieldsMap.get(name);
-                    RowType.Field column = new RowType.Field(columnName, columnType);
-                    fieldsBuilder.add(column);
-                    iterator.remove();
-                }
-            }
-            fieldsMap.put(prefix, RowType.from(fieldsBuilder.build()));
-        }
-        else {
-            boolean newColumnFound = columns.stream()
-                    .noneMatch(column -> column.getName().equalsIgnoreCase(field));
-            if (newColumnFound) {
-                columns.add(new ElasticsearchColumn(field, type, field, type.getDisplayName(), arrays.contains(field), -1));
-            }
-            fieldsMap.remove(field);
-        }
-        processNestedFields(fieldsMap, columns, arrays);
-    }
-
-    private List<String> getColumnMetadata(Optional<String> parent, JsonNode propertiesNode)
-    {
-        ImmutableList.Builder<String> metadata = ImmutableList.builder();
-        Iterator<Map.Entry<String, JsonNode>> iterator = propertiesNode.fields();
-        while (iterator.hasNext()) {
-            Map.Entry<String, JsonNode> entry = iterator.next();
-            String key = entry.getKey();
-            JsonNode value = entry.getValue();
-            String childKey;
-            if (parent.isPresent()) {
-                if (parent.get().isEmpty()) {
-                    childKey = key;
-                }
-                else {
-                    childKey = parent.get().concat(".").concat(key);
-                }
-            }
-            else {
-                childKey = key;
-            }
-
-            if (value.isObject()) {
-                metadata.addAll(getColumnMetadata(Optional.of(childKey), value));
-                continue;
-            }
-
-            if (!value.isArray()) {
-                metadata.add(childKey.concat(":").concat(value.textValue()));
-            }
-        }
-        return metadata.build();
-    }
-
-    private static Type getPrestoType(String elasticsearchType)
-    {
-        switch (elasticsearchType) {
-            case "double":
-            case "float":
-                return DOUBLE;
-            case "integer":
-                return INTEGER;
-            case "long":
-                return BIGINT;
-            case "string":
-            case "text":
-            case "keyword":
-                return VARCHAR;
-            case "boolean":
-                return BOOLEAN;
-            case "binary":
-                return VARBINARY;
-            default:
-                throw new IllegalArgumentException("Unsupported type: " + elasticsearchType);
-        }
-    }
-
-    private static class FieldNestingComparator
-            implements Comparator<String>
-    {
-        FieldNestingComparator()
-        {
-        }
-
-        @Override
-        public int compare(String left, String right)
-        {
-            // comparator based on levels of nesting
-            int leftLength = left.split("\\.").length;
-            int rightLength = right.split("\\.").length;
-            if (leftLength == rightLength) {
-                return left.compareTo(right);
-            }
-            return rightLength - leftLength;
-        }
     }
 }
